resolving conflicts: adding poi display handling; fixing exceptions handling
