+++ expo-maps/android/src/main/java/expo/modules/maps/googleMaps/GoogleMapsPlaces.kt
@@ -1,9 +1,7 @@
 package expo.modules.maps.googleMaps
 
 import android.content.Context
-import android.widget.Toast
 import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.model.PointOfInterest
 import com.google.android.libraries.places.api.*
 import com.google.android.libraries.places.api.net.*
 import expo.modules.kotlin.Promise
@@ -11,16 +9,15 @@ import expo.modules.maps.googleMaps.placesUtils.GoogleMapsPlacesSearchCompleter
 import expo.modules.maps.googleMaps.placesUtils.GoogleMapsPlacesTokenUtils
 import expo.modules.maps.googleMaps.placesUtils.GooglePlacesFetchPlace
 
-class GoogleMapsPlaces(private val context: Context, map: GoogleMap, private val markers: GoogleMapsMarkers): GoogleMap.OnPoiClickListener {
+class GoogleMapsPlaces(context: Context, map: GoogleMap, private val markers: GoogleMapsMarkers) {
 
     private val placesClient: PlacesClient
     private val tokenUtils: GoogleMapsPlacesTokenUtils
     private val placesSearchCompleter: GoogleMapsPlacesSearchCompleter
     private val placesFetcher: GooglePlacesFetchPlace
-    private var clickablePOIs: Boolean = true
 
     init {
-        Places.initialize(context, "apiKey")
+        Places.initialize(context, "GooglePlaces API Key")
         placesClient = Places.createClient(context)
         tokenUtils = GoogleMapsPlacesTokenUtils()
 
@@ -34,30 +31,12 @@ class GoogleMapsPlaces(private val context: Context, map: GoogleMap, private val
 
    fun createSearchRequest(place: String) {
        val placeId = getPlaceIdFromCompletion(place)
-       if (placeId.isBlank()) {
-           markers.detachAndDeletePOIMarkers()
-       }
-       placesFetcher.search(placeId)
+       if (placeId.isNotBlank()) placesFetcher.search(placeId) else markers.detachAndDeletePOIMarkers()
    }
 
     private fun getPlaceIdFromCompletion(place: String): String {
         val tmpStr = place.split(';')
         return if (tmpStr.size > 1) tmpStr[1] else ""
     }
-
-    fun setClickablePOIs(clickablePOIs: Boolean) {
-        this.clickablePOIs = clickablePOIs
-    }
-
-    override fun onPoiClick(poi: PointOfInterest) {
-        if (clickablePOIs) {
-            Toast.makeText(this.context, """Clicked: ${poi.name}
-            Place ID:${poi.placeId}
-            Latitude:${poi.latLng.latitude} Longitude:${poi.latLng.longitude}""",
-                    Toast.LENGTH_SHORT
-            ).show()
-        }
-    }
-
 
 }
